+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - CorrelationAnalysis
 - garmin.activity models
+- activities analysis
 
 ### Refactored
 - split GCDayTimeline in garmin.timeline and garmin.utils packages